Restructure implementation in C, so it's more modular
