+++ ext/normal_distribution/model.h
@@ -1,0 +1,9 @@
+#ifndef NORMAL_DISTRIBUTION_MODEL_H
+#define NORMAL_DISTRIBUTION_MODEL_H
+
+#include "ruby.h"
+#include "erf_inv.h"
+
+void Init_model( void );
+
+#endif //NORMAL_DISTRIBUTION_MODEL_H
